NabilahNordin (2225498) Remove Error and intergrate into one full code
